Move tiles calcluation into utils
