+++ src/utils.js
@@ -1,3 +1,5 @@
+import { point, rhumbBearing, rhumbDestination } from '@turf/turf'
+
 const d2r = Math.PI / 180
 
 const clip = (v, max) => {
@@ -200,4 +202,44 @@ export const getAdjustedOffset = (offset, renderedKey) => {
     Math.floor(offsetX / factor) + (renderedX % descendantFactor),
     Math.floor(offsetY / factor) + (renderedY % descendantFactor),
   ]
+}
+
+export const getTilesOfRegion = (region, level) => {
+  const { center, radius } = region.properties
+  const centralTile = pointToTile(center.lng, center.lat, level)
+
+  const tiles = new Set([tileToKey(centralTile)])
+
+  region.geometry.coordinates[0].forEach(([lng, lat]) => {
+    // Add tile along edge of region
+    const edgeTile = pointToTile(lng, lat, level)
+    tiles.add(tileToKey(edgeTile))
+
+    // Add any intermediate tiles if edge is > 1 tile away from center
+    const maxDiff = Math.max(
+      Math.abs(edgeTile[0] - centralTile[0]),
+      Math.abs(edgeTile[1] - centralTile[1])
+    )
+    if (maxDiff > 1) {
+      const centerPoint = point([center.lng, center.lat])
+      const bearing = rhumbBearing(centerPoint, point([lng, lat]))
+
+      for (let i = 1; i < maxDiff; i++) {
+        const intermediatePoint = rhumbDestination(
+          centerPoint,
+          (i * radius) / maxDiff,
+          bearing,
+          { units: 'miles' }
+        )
+        const intermediateTile = pointToTile(
+          intermediatePoint.geometry.coordinates[0],
+          intermediatePoint.geometry.coordinates[1],
+          level
+        )
+        tiles.add(tileToKey(intermediateTile))
+      }
+    }
+  })
+
+  return Array.from(tiles)
 }
